Move bucket related functions to its own file
